backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
Pillow==10.3.0
python-dotenv==1.0.1
